Added the new version of the library and moved some tests.
