+++ pom.xml
@@ -10,7 +10,7 @@
 	</parent>
 	<groupId>artie.web</groupId>
 	<artifactId>pedagogicalinterventionws</artifactId>
-	<version>0.15.0-RELEASE</version>
+	<version>0.16.0-RELEASE</version>
 	<name>pedagogical intervention webservice</name>
 	<description>ARTIE Pedagogical intervention webservice</description>
